imelunaservice: validate the LS2 payloads before acting on them
com.webos.service.ime is reachable by any client allowed on the bus, and the
numeric parameters went straight from JSON into int:

* deleteCharacters took "(int) characterCount.toDouble()" after testing only
  that the value was positive. A double that does not fit in an int makes that
  cast undefined behaviour, and the result was then used as a replace length.
* insertText did the same with replaceLength.

Both now go through intParameter(), which range-checks against the int the
callee expects and rejects NaN, and reply with an error instead of acting on a
garbage count.

registerRemoteKeyboard ignored whether LSSubscriptionAdd() succeeded: it
recorded the client in m_clientByToken and replied "subscribed": true either
way. With no subscription there is no cancel callback, so that entry stayed
forever, hasSubscribers() never went back to false, and the service kept
building and broadcasting widget state for a client that was not listening.
Report the failure and do not record the client.

insertText() and deleteCharacters() also used the cursor position from
surroundingText() without checking that the call succeeded. It returns false
with cursorPosition set to -1 when the widget has no surrounding text, and
negating that produced a negative replace start; deleteCharacters could also
reach sendCommitString() with a negative count when the cursor sat at the
start of the field. Both now bail out when there is no usable cursor position,
which also removes the INT_MIN special cases that were guarding the negation.

Two lifetime nits in the same file: LSErrorWrapper owns its LSError payload
but was copyable, so any copy would have freed it twice; and the destructor
unref'd the GMainLoop before LSUnregister(), tearing down the loop the handle
was still attached to.

+++ src/imelunaservice.h
@@ -56,7 +56,7 @@ protected:
     bool hasSubscribers() const;
 
     QJsonObject getWidgetStateJson() const;
-    void insertText(const QString& text, bool replace, ssize_t length = 0);
+    void insertText(const QString& text, bool replace, int length = 0);
     void deleteCharacters(int numChars, DeleteMode mode);
     void sendEnterKey();
